requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-debug-toolbar==4.2.0
django-filter==23.3
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
Markdown==3.5
packaging==23.2
phonenumbers==8.13.23
Pillow==10.1.0
psycopg-binary==3.1.12
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==68.2.2
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
wheel==0.41.3
